+++ packages/maker-experimental/src/conversion/toMarketplace/convertOntologyDefinitionToWireBlockData.ts
@@ -18,6 +18,7 @@ import type {
   ActionTypeBlockDataV2,
   ActionTypePermissionInformation,
   InterfaceTypeBlockDataV2,
+  KnownMarketplaceIdentifiers,
   LinkTypeBlockDataV2,
   ObjectTypeBlockDataV2,
   OntologyBlockDataV2,
@@ -139,7 +140,7 @@ export function convertOntologyDefinitionToWireBlockData(
 function buildKnownIdentifiers(
   ontology: OntologyDefinition,
   ridGenerator: OntologyRidGenerator,
-) {
+): KnownMarketplaceIdentifiers {
   // Interface types: InterfaceTypeRid -> BlockInternalId
   const interfaceMappings = Object.fromEntries(
     Array.from(ridGenerator.getInterfaceRids().inverse().entries()).map((
@@ -403,28 +404,28 @@ function buildKnownIdentifiers(
     actionParameterIds: actionParameterIdMappings,
     actionParameters: actionParameterMappings,
     actionTypes: actionTypeMappings,
-    datasourceColumns: datasourceColumns,
+    datasourceColumns,
     datasources: backingDatasourceMappings,
     filesDatasources: filesDatasourceMappings,
     functions: {},
-    geotimeSeriesSyncs: geotimeSeriesSyncs,
+    geotimeSeriesSyncs,
     groupIds: groupMappings,
     interfaceLinkTypes: interfaceLinkMappings,
     interfacePropertyTypes: interfacePropertyMappings,
     interfaceTypes: interfaceMappings,
-    linkTypeIds: linkTypeIds,
+    linkTypeIds,
     linkTypes: linkTypeRids,
     markings: markingsMappings,
-    objectTypeIds: objectTypeIds,
+    objectTypeIds,
     objectTypes: objectTypeRids,
-    propertyTypeIds: propertyTypeIds,
+    propertyTypeIds,
     propertyTypes: propertyRids,
     sharedPropertyTypes: sharedPropertyMappings,
     shapeIdForOntologyAllowSchemaMigrations: ridGenerator.toBlockInternalId(
       MIGRATION_SHAPE_READABLE_ID,
     ),
     shapeIdForInstallPrefix: null,
-    timeSeriesSyncs: timeSeriesSyncs,
+    timeSeriesSyncs,
     valueTypes: valueTypeMappings,
     webhooks: {},
     workshopModules: {},