+++ packages/create-app.template.expo.v2/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.expo.v2
 
+## 2.8.0-beta.30
+
+## 2.8.0-beta.29
+
 ## 2.8.0-beta.28
 
 ### Minor Changes