+++ packages/generator-converters.ontologyir/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/generator-converters.ontologyir
 
+## 2.8.0-beta.30
+
+### Patch Changes
+
+- @osdk/client.unstable@2.8.0-beta.30
+
+## 2.8.0-beta.29
+
+### Patch Changes
+
+- @osdk/client.unstable@2.8.0-beta.29
+
 ## 2.8.0-beta.28
 
 ### Minor Changes