+++ packages/react-components-styles/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @osdk/react-components-styles
 
+## 0.2.0-beta.27
+
+## 0.2.0-beta.26
+
+### Minor Changes
+
+- c3da451: add marking button, category group, and validation components for CBAC picker
+- 66b7125: Add RadioButtonsField component for action form radio button inputs
+
 ## 0.2.0-beta.25
 
 ## 0.2.0-beta.24