+++ packages/widget.client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/widget.client
 
+## 3.5.0-beta.30
+
+### Patch Changes
+
+- @osdk/widget.api@3.5.0-beta.30
+
+## 3.5.0-beta.29
+
+### Patch Changes
+
+- @osdk/widget.api@3.5.0-beta.29
+
 ## 3.5.0-beta.28
 
 ### Patch Changes